aspect_ratio强制转化double类型

+++ app/Models/v2/Video.php
@@ -92,7 +92,7 @@ class Video extends BaseModel
 
     public function getAspectRatioAttribute()
     {
-        return (double)$this->attributes['aspect_ratio'];
+        return (float)$this->attributes['aspect_ratio'];
     }
 
     public function getPromosAttribute()
